Valmistaja, jolla ei ole tuotteita voidaan ainoastaan poistaa nyt

+++ src/main/java/k25tiimi1backend/k25tiimi1backend/web/ManufacturerController.java
@@ -6,6 +6,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import k25tiimi1backend.k25tiimi1backend.domain.Manufacturer;
 import k25tiimi1backend.k25tiimi1backend.domain.ManufacturerRepository;
@@ -38,7 +39,21 @@ public class ManufacturerController {
     }
 
     @GetMapping("/deletemanufacturer/{id}")
-    public String deleteManufacturer(@PathVariable("id") Long manufacturerId) {
+    public String deleteManufacturer(@PathVariable("id") Long manufacturerId,  RedirectAttributes redirectAttributes) {
+    Manufacturer manufacturer = manufacturerRepository.findById(manufacturerId).orElse(null);
+
+    if (manufacturer == null) {
+        redirectAttributes.addFlashAttribute("error", "Manufacturer not found");
+        return "redirect:/manufacturerlist";
+    }
+
+    
+    if (manufacturer.getProducts() != null && !manufacturer.getProducts().isEmpty()) {
+        redirectAttributes.addFlashAttribute("error", "Cannot delete manufacturer with existing products :-( ");
+        return "redirect:/manufacturerlist";
+    }
+
+    
     manufacturerRepository.deleteById(manufacturerId);
     return "redirect:/manufacturerlist";
 }
